refactor #260518: メインレイアウト、コンポーネント分割
